Sally v1.0.14
Unicode path preservation through file ops, F-key launchers, and Find dialog,
plus Find dialog dark-mode polish and the review-comment fixes that followed.

Unicode path preservation
- Panel paths held as wide; REG_SZ Unicode persistence.
- Directory history with wide twins (deferred flush, decoupled serializer).
- Snooper wide variants via FindFirstChangeNotificationW.
- Main window registered Unicode-aware; worker source paths rebound.
- F5 copy/move: CFileData::NameW allocation broadened, BuildScriptFile
  lossy-leaf fallback hardened, pathW refreshed after ANSI path split,
  SourceNameWExplicit / TargetNameWExplicit COperation flags so
  PopulateWidePathsFromAnsi and ReanchorWideSourcePaths preserve
  explicitly-set wide names.
- F4 edit and Alt+F3 view routed through ::CreateProcessW with wide
  cmdLine + cwd; \$(FullPath)/\$(Name)/\$(Path) recovered via post-hoc
  lossy→wide substring substitution (both with- and without-trailing-
  backslash flavors of \$(Path)). Process/thread handles registered
  via HANDLES_ADD so the matching HANDLES(CloseHandle) finds them.
- Find dialog Unicode support: open-as-Unicode dialog, wide seed for
  "Look in", wide-aware combo subclass, post-Transfer override message.
  LookInTextW is authoritative only while the Unicode edit control is
  enabled — invariant enforced at read site, write site, and the
  constructor seed moved to the override handler.

Find dialog dark-mode polish
- Recover Find dialog borders / clean up borders.
- Skin Find combo edit rims; prevent Find combo theme recursion.
- Polish Find advanced controls; tone down Find separators.

Review-comment fixes rolled into the flatten
- regedt plugin x86 C2695: WINAPI + virtual...override on
  GetCurrentPathW to match the new CPluginFSInterfaceAbstract virtual.
- SalSplitWindowsPathW given a dedicated SAL_MAX_LONG_PATH wide buffer;
  wide mask re-converted via WideToAnsi into the ANSI path buffer past
  the null (mask no longer reads stale pre-splitter bytes; DBCS-safe).
- AlpDX wide-capacity tracking via CStaticText::AlpDXAllocated; both
  SetText (ANSI) and SetTextW realloc on demand (DBCS-safe).
- SalParsePathW preserves the split point: boundaryOff = end - buffer
  (correct for both existing-file break and new-dir loop).
- HDROP paste honors pastePath when set by the shell-extension
  context-menu Paste (previously pasted into the panel root).
- AskCannotMove ANSI args repacked to (fileName, destPath, errorText)
  matching the dialog handler and the wide sibling; AskCannotMoveErr
  wrapper updated to pass args in signature order.

Verification: sally Debug x64 builds clean; plugin_regedt Release x86
builds clean (-A Win32); 466 gtest cases pass across 12 suites.

Closes #56.

+++ src/common/dib.cpp
@@ -76,7 +76,7 @@ HBITMAP LoadBitmapAndMapColors(HINSTANCE hInst, HRSRC hRsrc, int mapCount,
     HDC hDCScreen = HANDLES(GetDC(NULL));
     LPBYTE lpBits;
     lpBits = (LPBYTE)(lpBitmap + 1);
-    lpBits += (1 << (lpBitmapInfo->biBitCount)) * sizeof(RGBQUAD);
+    lpBits += ((size_t)1 << lpBitmapInfo->biBitCount) * sizeof(RGBQUAD);
     HBITMAP hbm = HANDLES(CreateDIBitmap(hDCScreen, lpBitmapInfo, CBM_INIT, lpBits,
                                          (LPBITMAPINFO)lpBitmapInfo, DIB_RGB_COLORS));
     if (hbm == NULL)
